+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Service release addressing all zero-day issues reported so far...
 - fix diffusers dpm++ 2m and 1s samplers
 - fix new style filename template
 - fix image name template using model name
+- fix model path using relative path
 - force second requirements check on startup
 - remove lyco, multiple_tqdm
 - enhance extension compatibility for exensions directly importing codeformers